split large local
